- all quiz funtion: done - fix UI

+++ src/main/java/com/hanu/leaniverse/controller/QuestionController.java
@@ -2,9 +2,13 @@ package com.hanu.leaniverse.controller;
 
 import com.hanu.leaniverse.model.Question;
 import com.hanu.leaniverse.model.Quizz;
+import com.hanu.leaniverse.model.User;
 import com.hanu.leaniverse.repository.QuestionRepository;
 import com.hanu.leaniverse.repository.QuizzRepository;
+import com.hanu.leaniverse.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -18,13 +22,20 @@ public class QuestionController {
     @Autowired
     QuizzRepository quizzRepository;
 
-    @GetMapping("/show-question-management")
-    public String showAllQuestionEditPage(@RequestParam int quizzId, Model model){
-        List<Question> list = questionRepository.findQuestionsByQuizzId(quizzId);
+    @Autowired
+    UserService userService;
+
+    @GetMapping("/tutor/course/unit/quizz/{quizzId}")
+    public String showAllQuestionEditPage(@PathVariable("quizzId") int quizzId, Model model, Authentication authentication){
+        User user = userService.getCurrentUser(authentication);
+        if (user == null) {
+            return "redirect:/login";
+        }
+        Quizz quizz = quizzRepository.findById(quizzId).get();
         Question question = new Question();
-        model.addAttribute("questions",list);
+        model.addAttribute("quizz",quizz);
         model.addAttribute("questionForm",question);
-        return "questionManagePage";
+        return "/tutor/quizz";
     }
 
 //    @PostMapping("/addQuestion")
@@ -33,17 +44,26 @@ public class QuestionController {
 //        return "redirect:/showQuestionManagement";
 //    }
 
-    @PostMapping("/add-question")
-    public String addQuestion(@ModelAttribute Question question, @RequestParam int quizzId) {
+    @PostMapping("/tutor/course/unit/quizz/{quizzId}/add-question")
+    public String addQuestion(@ModelAttribute Question question, @PathVariable("quizzId") int quizzId, Authentication authentication) {
+        User user = userService.getCurrentUser(authentication);
+        if (user == null) {
+            return "redirect:/login";
+        }
+
         Quizz quizz = quizzRepository.findById(quizzId).orElseThrow(() -> new IllegalArgumentException("Invalid quizzId"));
         question.setQuizz(quizz);
         questionRepository.save(question);
-        return "redirect:/showQuestionManagement?quizzId=" + quizzId;
+        return "redirect:/tutor/course/unit/quizz/" + quizzId;
     }
 
-    @PostMapping("/delete-question")
-    public String deleteQuestion(@RequestParam int questionId, @RequestParam int quizzId){
+    @PostMapping("/tutor/course/unit/quizz/{quizzId}/delete-question/{questionId}")
+    public String deleteQuestion(@PathVariable int questionId, @PathVariable int quizzId, Authentication authentication){
+        User user = userService.getCurrentUser(authentication);
+        if (user == null) {
+            return "redirect:/login";
+        }
         questionRepository.deleteById(questionId);
-        return "redirect:/showQuestionManagement?quizzId=" + quizzId;
+        return "redirect:/tutor/course/unit/quizz/" + quizzId;
     }
 }
